+++ lib/widgets/viewer/info/basic_section.dart
@@ -13,7 +13,7 @@ import 'package:aves/utils/file_utils.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/common/identity/aves_filter_chip.dart';
 import 'package:aves/widgets/viewer/info/common.dart';
-import 'package:aves/widgets/viewer/info/entry_info_action_delegate.dart';
+import 'package:aves/widgets/viewer/action/entry_info_action_delegate.dart';
 import 'package:aves/widgets/viewer/info/owner.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';